refactor: Extract specs from Configuration
